Upgrade dummy devices
Summary:
Dummy devices are not "real devices" but rather devices that are created in the absence of them.

Sometimes, when trying to find a real device which a specific serial, it will fail, and a dummy device with said serial will be created instead.

Later on, it is possible for the real device to become available. In this case, it should be possible to promote the dummy device to an actual device.

Reviewed By: passy

Differential Revision: D57163208

fbshipit-source-id: 0d42d586ba767f5bee37af9a9d4c52b7518d4adc