Support parsing json bodys with numbers that are > 2^53
Summary:
Context
https://fb.workplace.com/groups/flipper.community.support/permalink/1005276087877573/

Since network plugin sends json bodys as raw json blobs we can deserialise it with a library that supports big int.

This allows us to show numbers in the way native clients see it.

This will not work for headers since they are send as part of the main flipper message which is parsed by the browser and does not support big int

whilst the added library carrys a perf impact we only pay the cost when  the user clicks and we render the sidebar

Reviewed By: passy

Differential Revision: D57780934

fbshipit-source-id: 622d3f99248d984538cbbd91da2f8c527e1cd5c8