Fix multiple fake devices during QR exchange
Summary:
Device UDID is not available during the exchange. As such, we do a best attempt to find devices by name. This is fine except we append "via QR exchange" to said name which then makes it impossible to match.

Displaying "via QR exchange" also exposes an implementation detail which does not have a direct relationship with the device. so removing this from the dropdown displayed title seems appropriate.

Reviewed By: antonk52

Differential Revision: D58415038

fbshipit-source-id: 1518627d58c0134b0ed2c5e981c85ca2a492b992