Prevent selected node from auto expanding when new frame come in
Summary:
addresses https://fb.workplace.com/groups/flippersupport/permalink/1864127004067949/

on receiving a frame we call a method  ensureAncestorsExpanded on the selected node.

If the user has collapsed the tree containing the selected node this causes it to re-expand on the next frame.

this is a bit janky given that when a frame comes in there might be no user interaction (videos playing / animations trigger new frames)

I cant think of a good reason this expansion needs to occur so opted to remove it

Reviewed By: antonk52

Differential Revision: D59163209

fbshipit-source-id: cf91401fa811d372ec85853e03c04bd4da4ad542