replace ts-node with tsx
Summary:
As part of simplifying codebase we can remove a custom command to run ts-node in favor of tsx(transforms typescript using esbuild instead of tsc).

Might be marginal speed difference which is irrelevant. Main reason is that it provides an easier way to run typescript code on all platforms.

We already use esbuild to bundle plugins so this fits nicely.

Reviewed By: mweststrate

Differential Revision: D59109777

fbshipit-source-id: 3dbf06552aadf0e2c5a873a6d73e1aa425800f02