Moved package from AndroidManifest.xml into build.gradle namespace (#5541)
Summary:
This is necessary because Google requires this in AGP version 8.

The code no longer works now that we're at version 8 of AGP. This is because AGP 8 requires that the package be changed to a namespace option in build.gradle.

## Changelog

The React Native module now works with Android Gradle Plugin version 8.

Pull Request resolved: https://github.com/facebook/flipper/pull/5541

Test Plan:
I imported this in my own React Native project, and I was getting a build failed: "Namespace not specified. Specify a namespace in the module's build file. See https://d.android.com/r/tools/upgrade-assistant/set-namespace for information about setting the namespace."

Then I made this one change and rebuilt my project. The build worked and my app ran correctly.

Reviewed By: LukeDefeo

Differential Revision: D55419610

Pulled By: passy

fbshipit-source-id: 5646172b14ab8a4ec6b6d3ffeb3b68c44e707883