Add videoId as a parameter to NetworkReporter
Summary:
Context
=
I am creating a flipper plugin to track and show network requests in a waterfall display to help IG App developers

See this [GSD](https://www.internalfb.com/gsd/316319693088236/791974585995443/list?t=189784262) for more context

FarAndHeight recommended a powersearching feature that woulda llow users to highlight individual requests.  dawidcieslak recommended a feature that would allow users to isolate media of a specific media-id. I plan to combine both of these features by storing the media-id of video requests and using the powersearch to allow for isolating requests that have a specific media id.

This Diff
=
This diff modifies `NetworkReporter.java`, which contains the type information used for request information sent to the desktop, to include a videoId parameter. If no videoId is passed in, the videoId parameter is defaulted to null.

Differential Revision: D59925835

fbshipit-source-id: 98acc56b31d495d246db3bef13b3b45a09f9c882