remove deprecation warning for 2 antd components
Summary:
There warnings are not actionable and pollute test output. Lets comment these out

* antd about deprecated component api (removes ~7k lines)
* react-dom about mounting app (removes ~2.5k lines)
* a few flipper logs that are commonly used during many test startup (~500 lines)

Number of lines that is outputed by `yarn run test`
```
Before: 10139
After:   1065
```

**New output**
{P1486339650}

There are still some console logs that are present in the output which are less common.

The reason I didn't add process.env.NODE_ENV check in our logger itself is that then people won't be able to use console.XXX methods in tests to inspect values which would be a nightmare to debug.

Reviewed By: lblasa

Differential Revision: D59962243

fbshipit-source-id: d8096b21deeff555d0025dd1e8814b745283a55c