NetworkFlipperPlugin#<init>'s param 'formatters' Receives Null Value
Summary:
mark formatters nullsafe as stated by the error code:

```
07-23 16:21:27.445  4845  4845 E AndroidRuntime: Caused by: java.lang.IllegalArgumentException: NetworkFlipperPlugin#<init>'s param 'formatters' received null value, but is not marked as Nullable.
07-23 16:21:27.445  4845  4845 E AndroidRuntime: 	at com.facebook.flipper.plugins.network.NetworkFlipperPlugin.<init>(NetworkFlipperPlugin.java:1)
07-23 16:21:27.445  4845  4845 E AndroidRuntime: 	at com.facebook.flipper.plugins.network.NetworkFlipperPlugin.<init>(NetworkFlipperPlugin.java:27)
07-23 16:21:27.445  4845  4845 E AndroidRuntime: 	at com.instagram.appinitializer.devtools.FlipperInitializer.init(FlipperInitializer.java:96)
07-23 16:21:27.445  4845  4845 E AndroidRuntime: 	at com.instagram.appinitializer.devtools.DevToolInitializer.init(DevToolInitializer.kt:27)
```

Also add nullable to field set to null + method returning null explicitly

## Context

We are in the process of making the entire codebase Nullsafe. This diff relies on a runtime codemod to track when null is passed into a parameter that is not marked as Nullable in java. Based on runtime data, we can be 100% certain that this parameter needs to be Nullable

Reviewed By: nicholeic

Differential Revision: D60174408

fbshipit-source-id: 0e41badfa743445bb295af25da30faaffcdc5477