update no devices found tooltip
Summary:
there was a tooltip for no device found but it was never in use since its icon was always hidden.

I've removed this separate icon and added the tooltip to the regular no devices found label and updated the links and styling.

Addionally i added a border and subtile animation to the no devices found div, this is to encourage users to hover it to see the tooltip. i think the slight annoyance of the animation is worth it as when you have no devices flipper is useless

Reviewed By: antonk52

Differential Revision: D60178471

fbshipit-source-id: 7c613c3736e429d7406f632d743704e23e639804