Don't Store Image Response bodies for Network and Network Chart plugin if there is no flipper connection
Summary:
Context
=
I am creating a flipper plugin to track and show network requests in a waterfall display to help IG App developers

See this [GSD](https://www.internalfb.com/gsd/316319693088236/791974585995443/list?t=189784262) for more context

This Diff
=
As the title indicates, this diff strips response bodies if there's no flipper connection in Network and IGNetworkingBehavior plugins. This is to reduce the amount of memory used by the plugin and prevent OOM issues for debug builds for developers not using flipper.

Reviewed By: pfarcasanu

Differential Revision: D60175106

fbshipit-source-id: 07d1895b6da15fdca70719606f4ef46ef0b02688