Update ApplicationWrapper param that received null value to Nullable
Summary:
Runtime plugin showed null savedInstanceState passed into onActivityCreated.

## Context
We are in the process of making the entire codebase Nullsafe. This diff relies on a runtime codemod to track when null is passed into a parameter that is not marked as Nullable in java. Based on runtime data, we can be 100% certain that this parameter needs to be Nullable

Reviewed By: jocelynluizzi13

Differential Revision: D60469755

fbshipit-source-id: 2d4385326bff7c4f75c1c26afd2552d2be4fa42b