Update SharedPreferencesFlipperPlugin param that received null value to Nullable
Summary:
Runtime plugin showed null key passed into onSharedPreferenceChanged. Required buck dependency to allow nullable annotation.

## Context
We are in the process of making the entire codebase Nullsafe. This diff relies on a runtime codemod to track when null is passed into a parameter that is not marked as Nullable in java. Based on runtime data, we can be 100% certain that this parameter needs to be Nullable

Reviewed By: jocelynluizzi13

Differential Revision: D60586640

fbshipit-source-id: 1740e3bbec73186c908ed2cc31f52c0cd5fc12d9