Use source maps instead of debug metadata for components 1/n
Summary:
The code is a lot simpler and more readable now because the source maps only contain the actual derived component chain that produced a bound tree node rather than all the way up to the root.

We use an LRU cache to store a set number of derived component chains for a certain number of bound tree nodes.

Requests are done in batches and grouped on source map id because some ids are not stored since they are prepacked, this causes teh entire request to fail

the network call doesnt contain nested data, this will be fixed in a future diff once the server side is ready

Differential Revision: D59912728

fbshipit-source-id: 4f22dc174d8ed97f249b01ed442c22e40e1954d4