Creating C++ FlipperMultiPlugin
Summary:
Currently, flipper does not support to have multiple plugins with the same identifiers. For example, there are 2 Flipper QPL plugins - one on Java and another on C++ sides with the same "UIPerf" identifier. Whatever call was made first from these 2 plugins will be added to the client. Changing this by adding a new "FlipperMultiPlugin" definition - that will contain several plugins.

This will be used to add support of the Flipper Plugin to the native QPL in the next diffs

Reviewed By: mkareta

Differential Revision: D63095017

fbshipit-source-id: b991a61f6deba4c627a214224c3fceabbe274a49