Modifying addPlugin() based on the isDuplicationAllowed flag
Summary:
Adding the flag if the FlipperPlugin is allowed for duplication. This will be set as False by default.

Since there is already existing Multi Plugin and it is possible to find if the current plugin allowed for duplication, modifying the addPlugin() function. The behaviour will stay the same for all plugins that has the newly added flag by default.

If the plugin with the same identifier is already present and allowed for duplication, then instead of skipping - will create an instance of the multi plugin. Also cleaning the already existing one by disconnecting and erasing it.

The newly changed logic can still cause the tree-shape structure, however this might be possible only for the users that will modify the flag and call the addPlugin() few times with the same identifier. As for now, I presume this might be not a problem

Differential Revision: D63325554

fbshipit-source-id: affbc9f4bac1f71c37c5f284e707a324298031f5