Fixing the hybrid qpl integration with the flipper plugin
Summary:
In the previous diff hybrid qpl data is started being missing due to JFlipperPlugin not having the isDuplicationAllowed field. Because FlipperPlugin is implemented as the Java Interface it wasn't possible to add it there - instead we added this field to the QPL Flipper Plugin.

This is causing an issue when the C++ flipper plugin will be added first to the client. Because isDuplicationAllowed flag as the default will be False. For this reason, Java Flipper plugin is not being attached when calling the native addPlugin() API. Changing this call to check for the identifier and if it's "UIPerft"(Flipper plugin) then changing it to True.

Differential Revision: D64058286

fbshipit-source-id: b74810ba9c4e91956545ae4fd3260fc4159ff7c1