Client error message
Summary:
See previous diff in stack for context, serialization can fail and when it does we just send an empty string payload to flipper which is v confusing.

Instead we now do proper error handling

A note about the use of try catch around nsjson serialization i know this is not typical but the previous approach was to use the validator function to see if data could be serialized. this was returning false before but the problem is we have idea what the error is. Unfortunatley despite taking an error parm the NSJSONSerialization dataWithJSONObject function throws an exception so im force to catch it to get a useful error message

Differential Revision: D65605276

fbshipit-source-id: 506ecc48bd0692a622679a8f621fede7d626c122