Don't leak CRYPTO objects in generateCertPKCS12
Summary:
`bioFromFile` is still allocating a BIO object which is not free'd.

I'm trying to chase a memory leak in media creation in IG therefore I need a local build with malloc tracing enabled and this leak it's introducing noise in my leaks tool analysis.

(I tried to disable flipper for local builds but it's more difficult than actually fixing this leak)

Reviewed By: fabiomassimo

Differential Revision: D65686212

fbshipit-source-id: 5d97dc37815ce15f09aeef554c9049c47234233f